refactor: move header components to package
